requirements.txt
requests==2.25.1
beautifulsoup4==4.9.3
google-api-python-client==2.19.0
aiohttp==3.7.4
python-dotenv==0.17.1